requirements.txt
django
django-rest-framework
psycopg2-binary==2.9.3
gpt4all

huggingface-hub
sentence-transformers

langchain==0.0.352
chromadb==0.3.29
tiktoken
unstructured
unstructured[local-pdf]
unstructured[local-inference]

llama-index
llama-index-llms-ollama
llama-index-embeddings-huggingface